Module rename and added comments
