+++ src/sickle/modules/diff.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-from sickle.common.lib.generic.mparser import argument_check
+from sickle.common.lib.generic.modparser import argument_check
 from sickle.common.lib.generic.extract import read_bytes_from_file
 
 from sickle.common.lib.generic.colors import Colors
