+++ src/sickle/payloads/linux/aarch64/shell_reverse_tcp.py
@@ -1,7 +1,7 @@
 import sys
 
 from sickle.common.lib.reversing.assembler import Assembler
-from sickle.common.lib.generic.mparser import argument_check
+from sickle.common.lib.generic.modparser import argument_check
 from sickle.common.lib.generic.convert import port_str_to_htons
 from sickle.common.lib.generic.convert import ip_str_to_inet_addr
