+++ src/sickle/payloads/linux/x64/memfd_reflective_elf_tcp.py
@@ -4,7 +4,7 @@ import struct
 
 import sickle.common.lib.generic.extract as extract
 import sickle.common.lib.generic.convert as convert
-import sickle.common.lib.generic.mparser as modparser
+import sickle.common.lib.generic.modparser as modparser
 import sickle.common.lib.programmer.builder as builder
 
 from sickle.common.lib.reversing.assembler import Assembler
