+++ src/sickle/payloads/windows/aarch64/shell_reverse_tcp.py
@@ -3,7 +3,7 @@ import ctypes
 import struct
 
 import sickle.common.lib.generic.convert as convert
-import sickle.common.lib.generic.mparser as modparser
+import sickle.common.lib.generic.modparser as modparser
 import sickle.common.lib.programmer.builder as builder
 
 from sickle.common.lib.reversing.assembler import Assembler
