+++ src/sickle/payloads/windows/x64/reflective_pe_tcp.py
@@ -5,7 +5,7 @@ import struct
 
 import sickle.common.lib.generic.extract as extract
 import sickle.common.lib.generic.convert as convert
-import sickle.common.lib.generic.mparser as modparser
+import sickle.common.lib.generic.modparser as modparser
 import sickle.common.lib.programmer.builder as builder
 
 from sickle.common.lib.reversing.assembler import Assembler
